+++ common-functional/src/main/java/cn/featherfly/common/function/serializable/SerializableShortSupplier.java
@@ -3,15 +3,15 @@ package cn.featherfly.common.function.serializable;
 
 import java.io.Serializable;
 
-import cn.featherfly.common.function.ByteSupplier;
+import cn.featherfly.common.function.ShortSupplier;
 
 /**
  * The Interface SerializableByteSupplier.
  *
  * @author zhongj
- * @see ByteSupplier
+ * @see ShortSupplier
  */
 @FunctionalInterface
-public interface SerializableShortSupplier extends Serializable, ByteSupplier {
+public interface SerializableShortSupplier extends Serializable, ShortSupplier {
 
 }
